(GH-3693) Moved the Attached HeaderProperties under HeaderedControlHelper for clarity
