+++ src/MahApps.Metro.Samples/MahApps.Metro.Demo/ExampleWindows/CleanWindowDemo.xaml
@@ -121,7 +121,7 @@
 
             <GroupBox Grid.Column="0"
                       Padding="5"
-                      Controls:ControlsHelper.HeaderFontSize="30"
+                      Controls:HeaderedControlHelper.HeaderFontSize="30"
                       Header="artists">
                 <ListBox x:Name="artist"
                          ItemsSource="{x:Static models:SampleData.Artists}"
@@ -170,7 +170,7 @@
 
             <GroupBox Grid.Column="2"
                       Padding="5"
-                      Controls:ControlsHelper.HeaderFontSize="30"
+                      Controls:HeaderedControlHelper.HeaderFontSize="30"
                       Header="some tests">
                 <StackPanel MinWidth="100"
                             HorizontalAlignment="Center"
