+++ src/MahApps.Metro.Samples/MahApps.Metro.Demo/ExampleWindows/FlyoutDemo.xaml
@@ -95,7 +95,7 @@
             <Controls:Flyout x:Name="settingsFlyout"
                              Margin="200 30 0 0"
                              Controls:ControlsHelper.ContentCharacterCasing="Upper"
-                             Controls:ControlsHelper.HeaderFontFamily="Segoe Script"
+                             Controls:HeaderedControlHelper.HeaderFontFamily="Segoe Script"
                              AreAnimationsEnabled="True"
                              FocusedElement="{Binding ElementName=firstTB}"
                              Header="Settings"
