+++ src/MahApps.Metro/Controls/Helper/ControlsHelper.cs
@@ -67,89 +67,7 @@ namespace MahApps.Metro.Controls
             element.SetValue(ContentCharacterCasingProperty, value);
         }
 
-        public static readonly DependencyProperty HeaderFontFamilyProperty =
-            DependencyProperty.RegisterAttached("HeaderFontFamily", typeof(FontFamily), typeof(ControlsHelper), new FrameworkPropertyMetadata(SystemFonts.MessageFontFamily));
 
-        [Category(AppName.MahApps)]
-        [AttachedPropertyBrowsableForType(typeof(HeaderedContentControl))]
-        [AttachedPropertyBrowsableForType(typeof(TabControl))]
-        [AttachedPropertyBrowsableForType(typeof(Flyout))]
-        public static FontFamily GetHeaderFontFamily(UIElement element)
-        {
-            return (FontFamily)element.GetValue(HeaderFontFamilyProperty);
-        }
-
-        public static void SetHeaderFontFamily(UIElement element, FontFamily value)
-        {
-            element.SetValue(HeaderFontFamilyProperty, value);
-        }
-
-        public static readonly DependencyProperty HeaderFontSizeProperty =
-            DependencyProperty.RegisterAttached("HeaderFontSize", typeof(double), typeof(ControlsHelper), new FrameworkPropertyMetadata(SystemFonts.MessageFontSize));
-
-        [Category(AppName.MahApps)]
-        [AttachedPropertyBrowsableForType(typeof(HeaderedContentControl))]
-        [AttachedPropertyBrowsableForType(typeof(TabControl))]
-        [AttachedPropertyBrowsableForType(typeof(Flyout))]
-        public static double GetHeaderFontSize(UIElement element)
-        {
-            return (double)element.GetValue(HeaderFontSizeProperty);
-        }
-
-        public static void SetHeaderFontSize(UIElement element, double value)
-        {
-            element.SetValue(HeaderFontSizeProperty, value);
-        }
-
-        public static readonly DependencyProperty HeaderFontStretchProperty =
-            DependencyProperty.RegisterAttached("HeaderFontStretch", typeof(FontStretch), typeof(ControlsHelper), new FrameworkPropertyMetadata(TextElement.FontStretchProperty.DefaultMetadata.DefaultValue));
-
-        [Category(AppName.MahApps)]
-        [AttachedPropertyBrowsableForType(typeof(HeaderedContentControl))]
-        [AttachedPropertyBrowsableForType(typeof(TabControl))]
-        [AttachedPropertyBrowsableForType(typeof(Flyout))]
-        public static FontStretch GetHeaderFontStretch(UIElement element)
-        {
-            return (FontStretch)element.GetValue(HeaderFontStretchProperty);
-        }
-
-        public static void SetHeaderFontStretch(UIElement element, FontStretch value)
-        {
-            element.SetValue(HeaderFontStretchProperty, value);
-        }
-
-        public static readonly DependencyProperty HeaderFontWeightProperty =
-            DependencyProperty.RegisterAttached("HeaderFontWeight", typeof(FontWeight), typeof(ControlsHelper), new FrameworkPropertyMetadata(SystemFonts.MessageFontWeight));
-
-        [Category(AppName.MahApps)]
-        [AttachedPropertyBrowsableForType(typeof(HeaderedContentControl))]
-        [AttachedPropertyBrowsableForType(typeof(TabControl))]
-        [AttachedPropertyBrowsableForType(typeof(Flyout))]
-        public static FontWeight GetHeaderFontWeight(UIElement element)
-        {
-            return (FontWeight)element.GetValue(HeaderFontWeightProperty);
-        }
-
-        public static void SetHeaderFontWeight(UIElement element, FontWeight value)
-        {
-            element.SetValue(HeaderFontWeightProperty, value);
-        }
-
-        public static readonly DependencyProperty HeaderMarginProperty =
-            DependencyProperty.RegisterAttached("HeaderMargin", typeof(Thickness), typeof(ControlsHelper), new UIPropertyMetadata(new Thickness()));
-
-        [Category(AppName.MahApps)]
-        [AttachedPropertyBrowsableForType(typeof(HeaderedContentControl))]
-        [AttachedPropertyBrowsableForType(typeof(Flyout))]
-        public static Thickness GetHeaderMargin(UIElement element)
-        {
-            return (Thickness)element.GetValue(HeaderMarginProperty);
-        }
-
-        public static void SetHeaderMargin(UIElement element, Thickness value)
-        {
-            element.SetValue(HeaderMarginProperty, value);
-        }
 
         public static readonly DependencyProperty FocusBorderBrushProperty
             = DependencyProperty.RegisterAttached("FocusBorderBrush",
