+++ src/MahApps.Metro/Themes/MetroTabItem.xaml
@@ -48,17 +48,17 @@
                                                            ContentStringFormat="{TemplateBinding HeaderStringFormat}"
                                                            ContentTemplate="{TemplateBinding HeaderTemplate}"
                                                            ContentTemplateSelector="{TemplateBinding HeaderTemplateSelector}"
-                                                           FontFamily="{TemplateBinding Controls:ControlsHelper.HeaderFontFamily}"
-                                                           FontSize="{TemplateBinding Controls:ControlsHelper.HeaderFontSize}"
-                                                           FontStretch="{TemplateBinding Controls:ControlsHelper.HeaderFontStretch}"
+                                                           FontFamily="{TemplateBinding Controls:HeaderedControlHelper.HeaderFontFamily}"
+                                                           FontSize="{TemplateBinding Controls:HeaderedControlHelper.HeaderFontSize}"
+                                                           FontStretch="{TemplateBinding Controls:HeaderedControlHelper.HeaderFontStretch}"
                                                            FontStyle="{TemplateBinding FontStyle}"
-                                                           FontWeight="{TemplateBinding Controls:ControlsHelper.HeaderFontWeight}"
+                                                           FontWeight="{TemplateBinding Controls:HeaderedControlHelper.HeaderFontWeight}"
                                                            Foreground="{TemplateBinding Foreground}"
                                                            RecognizesAccessKey="True"
                                                            SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}" />
                                 <Button x:Name="PART_CloseButton"
                                         Grid.Column="1"
-                                        Width="{TemplateBinding Controls:ControlsHelper.HeaderFontSize, Converter={StaticResource MetroTabItemCloseButtonWidthConverter}}"
+                                        Width="{TemplateBinding Controls:HeaderedControlHelper.HeaderFontSize, Converter={StaticResource MetroTabItemCloseButtonWidthConverter}}"
                                         Height="{Binding RelativeSource={RelativeSource Self}, Path=Width, Mode=OneWay}"
                                         Margin="{TemplateBinding CloseButtonMargin}"
                                         VerticalAlignment="Top"
